Backend/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
flask-jwt-extended
bcrypt
werkzeug
psutil
